+++ docs/conf.py
@@ -1,9 +1,8 @@
-from __future__ import unicode_literals
-
 import os
 import re
 import subprocess
 import sys
+
 
 sys.path.append(os.path.abspath(".."))